fix: block through BankKeeper
