+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Fixed
 
 - Reject `MsgCreateVestingAccount`, `MsgCreatePeriodicVestingAccount`, and `MsgCreatePermanentLockedAccount` in the Cosmos ante (top-level and nested in `authz.MsgExec`) so new vesting / locked accounts cannot be opened after the v7.3.2 upgrade
-- Block the 22 Aug 2026 incident addresses in Cosmos/EVM ante (and strip/reject them in Prepare/ProcessProposal) so they cannot send, deploy, or be called after the v7.3.2 restart
+- Block the 22 Aug 2026 incident addresses with a bank `SendRestriction` so Cosmos, precompile, and EVM native transfers cannot send from or to them. Prepare/ProcessProposal still strip or reject txs that touch the list.
 
 
 ### Dependencies
